+++ comic_dl/__main__.py
@@ -2,6 +2,8 @@
 # -*- coding: utf-8 -*-
 
 from comic_dl import ComicDL
+import sites
+import manga_eden
 # import comic_dl
 import sys