+++ comic_dl/manga_eden/__init__.py
@@ -1,0 +1,3 @@
+import mangaSearch
+import mangaChapters
+import mangaChapterDownload